fix partialrenderer variants
