+++ packages/react-dom/src/server/ReactDOMLegacyServerNode.classic.fb.js
@@ -1,0 +1,19 @@
+/**
+ * Copyright (c) Facebook, Inc. and its affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ * @flow
+ */
+
+export {
+  renderToString,
+  renderToStaticMarkup,
+  version,
+} from './ReactDOMServerLegacyPartialRendererBrowser';
+
+export {
+  renderToNodeStream,
+  renderToStaticNodeStream,
+} from './ReactDOMLegacyServerNodeStream';
